Change jsoup tests related to resolving of URI.
In C# symbol "/" is added if absent to the URI's path, while in Java it stays the same as in the input string. Though both variants are correct the changes were made so that tests work in the same way for C# and Java. Also there are some differences in resolving file URIs.

DEVSIX-902