Normalize css properties on creation. Small fix in background shorthand resolver.
All properties are made lovercase with double spaces removed. In case space is followed by coma or left bracket it will be removed completely. In case space is after coma or right bracket it will be removed as well. This formating is not applied to text in single or double qoutes. Default font resolver was improved, now it can proceed lowercase default font-families.
Fix small bug in background shorthand resolver. The url with slash in it wasn't proceeded correctly.
Add new tests and fix corresponding ones.

DEVSIX-915