Fix large tables processing. Fix footer processing. Update cmps.
Large tables changes:
Do not process the bottom border with the corresponding row. This means we will collapse the bottom border of the last flushed row
and the top border of current row. If the table is splitting, process the bottom border alone.

Tables changes:
Do not collapse the bottom border with the next row top border if current row cannot be fully placed.
Fix FORCED_PLACEMENT and footer processing
Some fixed in row height processing.