Fix css pseudo class argument bug
Previously in some cases arguments weren't added to css selector
because of bug in pattern. Now pattern don't look for arguments but
they are checked in method that resolves css pseudo elements.

The new class is intended to unify C# and Java matcher behavior for
CssSelectorParser. It is supposed to be manually ported. The method
next(int) can be implemented in C# by storing offset and pattern and
recreating match with source substring.

DEVSIX-1309