Fix handling of initial value for border-color properties
Values 'initial' and 'inherit' are not allowed to be a
part of the shorthand properties values, they are
either the whole property value (e.g. border: inherit) or
shall not be used at all. CSS declarations like
'border: 2px solid initial" are invalid and should be
skipped.

Default border for <input> is currently set to `black` in
order to maintain current behaviour for now. It will be
fixed separately in future.

The 'outline-color' property default value changed to
'currentColor'. Accoring to specification for : "If the
user agent does not support the invert value then it
must reject that value at parse-time, and the initial value
of the outline-color property is the currentColor keyword".
Rejecting 'invert' value at parse-time is not supported for
now, instead `invert` behaves the same way as 'initial'.

DEVSIX-1562