Fix issues with text alignment inside select fields
The InfiniteWidthParagraphRenderer class was used in order to work around
not supporting of `nowrap` value for `white-space` css property. However
layout on infinite width breaks at least text alignment, and it seems that
it's more important to support text-alignment than some of the edge cases
of select fields with small width.

The `white-space: nowrap` will be supported in DEVSIX-1896.

DEVSIX-984