Add OWASP dependency check
Generates reports that are consumed by SonarQube, does not fail the build by itself

The OWASP Dependency Check is bound to the verify phase and runs in the Run Tests and Static Code Analysis stages. To speed up the pipeline, the OWASP Dependency Check doesn't run when the QA profile is used.

The changes in pom.xml override the values in the root pom (iText Core pom).

A future optimization of the pipeline is to have all static code analysis _before_ the Run Test phase, and have them produce their XML reports which can then be consumed by SonarQube

QA-6408 QA-8167 QA-8597