Remove FontSelectorPerformanceTest
Except for system fonts usage, the usecases mentioned in the tests are covered by other tests.
System fonts, however, could not be covered in a module checked on CI, while the installed font set
differs on each machine. The relevant test will be improved and moved to manual-tests repo.

DEVSIX-3687